requirements.txt
flask
flask_cors
python-dotenv
pandas
langchain
openai
faiss-cpu